v3 commit

+++ src/services/postgres/CollaborationsService.js
@@ -5,14 +5,15 @@ const InvariantError = require('../../exceptions/InvariantError');
 const nanoid = customAlphabet('1234567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', 16);
 
 class CollaborationsService {
-  constructor() {
+  constructor(cacheService) {
     this._pool = new Pool();
+    this._cacheService = cacheService;
   }
 
   async addCollaboration(userId, playlistId) {
     const collabId = `collab-${nanoid()}`;
     const query = {
-      text: 'INSERT INTO collaborations VALUES($1, $2, $3) RETURNING id',
+      text: 'INSERT INTO collaborations VALUES($1, $2, $3) RETURNING id, user_id',
       values: [collabId, userId, playlistId],
     };
 
@@ -21,7 +22,9 @@ class CollaborationsService {
     if (!result.rowCount) {
       throw new InvariantError('Kolaborasi berhasil ditambahkan');
     }
-
+    const { user_id: owner } = result.rows[0];
+    await this._cacheService.del(`playlists:${owner}`);
+    await this._cacheService.del(`playlist:${playlistId}`);
     return result.rows[0].id;
   }
 
@@ -39,7 +42,7 @@ class CollaborationsService {
 
   async deleteCollaboration(userId, playlistId) {
     const query = {
-      text: 'DELETE FROM collaborations WHERE user_id = $1 AND playlist_id = $2 RETURNING id',
+      text: 'DELETE FROM collaborations WHERE user_id = $1 AND playlist_id = $2 RETURNING id, user_id',
       values: [userId, playlistId],
     };
 
@@ -48,6 +51,10 @@ class CollaborationsService {
     if (!result.rowCount) {
       throw new InvariantError('Kolaborasi gagal dihapus');
     }
+
+    const { user_id: owner } = result.rows[0];
+    await this._cacheService.del(`playlists:${owner}`);
+    await this._cacheService.del(`playlist:${playlistId}`);
   }
 }
